// bindings/typescript/types/MikanMathTypes.ts
export class MikanQuatd {
  w: number = 0;
  x: number = 0;
  y: number = 0;
  z: number = 0;

  static __serializationMetadata: SerializationField[] = [
    { name: 'w', type: 'double' },
    { name: 'x', type: 'double' },
    { name: 'y', type: 'double' },
    { name: 'z', type: 'double' }
  ];
}
// ...
export class MikanVector3d {
  x: number = 0;
  y: number = 0;
  z: number = 0;

  static __serializationMetadata: SerializationField[] = [
    { name: 'x', type: 'double' },
    { name: 'y', type: 'double' },
    { name: 'z', type: 'double' }
  ];
}